# src/test/resources/openapi/test-api.yaml
openapi: 3.0.3
info:
  title: Spec for api tests.
  version: '1'
paths:
...
  /response/optional:
    get:
      operationId: optional
      tags: [response]
      parameters:
        - name: found
          in: query
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                type: string
        404:
          description: Not found.
...
  /param/path/requiredWithoutDefault/{string}:
    get:
      operationId: paramPathRequiredWithoutDefault
      tags: [parameter]
      parameters:
        - name: string
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/path/requiredWithDefault/{string}:
    get:
      operationId: paramPathRequiredWithDefault
      tags: [parameter]
      parameters:
        - name: string
          in: path
          required: true
          schema:
            type: string
            default: "this-is-default"
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/path/with-max-length/{string}:
    get:
      operationId: paramPathWithMaxLength
      tags: [parameter]
      parameters:
        - name: string
          in: path
          required: true
          schema:
            type: string
            maxLength: 5
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/path/with-pattern/{string}:
    get:
      operationId: paramPathWithPattern
      tags: [parameter]
      parameters:
        - name: string
          in: path
          required: true
          schema:
            type: string
            pattern: "[0-9][a-z]"
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/path/with-uuid/{uuid}:
    get:
      operationId: paramPathWithUuid
      tags: [parameter]
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/path/with-integer/{integer}:
    get:
      operationId: paramPathWithInteger
      tags: [parameter]
      parameters:
        - name: integer
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/path/with-long/{long}:
    get:
      operationId: paramPathWithLong
      tags: [parameter]
      parameters:
        - name: long
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/path/with-bool/{bool}:
    get:
      operationId: paramPathWithBoolean
      tags: [parameter]
      parameters:
        - name: bool
          in: path
          required: true
          schema:
            type: boolean
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/path/overlay/{string}:
    get:
      operationId: paramPathOverlayWithString
      tags: [parameter]
      parameters:
        - name: string
          in: path
          required: true
          schema:
            type: string
            maxLength: 10
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/path/overlay/{uuid}:
    get:
      operationId: paramPathOverlayWithUuid
      tags: [parameter]
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/query/multiple-lists:
    get:
      operationId: paramQueryMultipleLists
      tags: [parameter]
      parameters:
        - name: a
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: b
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
              format: uuid
        - name: c
          in: query
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int32
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/query/date:
    get:
      operationId: paramQueryDate
      tags: [parameter]
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/query/date-time:
    get:
      operationId: paramQueryDateTime
      tags: [parameter]
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/query/number:
    get:
      operationId: paramQueryNumber
      tags: [parameter]
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: number
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/query/uuid:
    get:
      operationId: paramQueryUuid
      tags: [parameter]
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/query/requiredWithoutDefault:
    get:
      operationId: paramQueryRequiredWithoutDefault
      tags: [parameter]
      parameters:
        - name: string
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
        400:
          description: Parameter missing.
  /param/query/requiredWithDefault:
    get:
      operationId: paramQueryRequiredWithDefault
      tags: [parameter]
      parameters:
        - name: string
          in: query
          required: true
          schema:
            type: string
            default: "this-is-default"
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/query/optionalWithoutDefault:
    get:
      operationId: paramQueryOptionalWithoutDefault
      tags: [parameter]
      parameters:
        - name: string
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/query/optionalWithDefault:
    get:
      operationId: paramQueryOptionalWithDefault
      tags: [parameter]
      parameters:
        - name: string
          in: query
          required: false
          schema:
            type: string
            default: "this-is-default"
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/header/requiredWithoutDefault:
    get:
      operationId: paramHeaderRequiredWithoutDefault
      tags: [parameter]
      parameters:
        - name: string
          in: header
          required: true
          schema:
            type: string
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
        400:
          description: Parameter missing.
  /param/header/requiredWithDefault:
    get:
      operationId: paramHeaderRequiredWithDefault
      tags: [parameter]
      parameters:
        - name: string
          in: query
          required: true
          schema:
            type: string
            default: "this-is-default"
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/header/optionalWithoutDefault:
    get:
      operationId: paramHeaderOptionalWithoutDefault
      tags: [parameter]
      parameters:
        - name: string
          in: header
          required: false
          schema:
            type: string
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/header/optionalWithDefault:
    get:
      operationId: paramHeaderOptionalWithDefault
      tags: [parameter]
      parameters:
        - name: string
          in: header
          required: false
          schema:
            type: string
            default: "this-is-default"
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/body/required:
    post:
      operationId: paramBodyRequired
      tags: [parameter]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParameterModel'
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
        400:
          description: Parameter missing.
  /param/body/optional:
    post:
      operationId: paramBodyOptional
      tags: [parameter]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParameterModel'
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
  /param/all/{path}:
    get:
      operationId: paramAll
      tags: [parameter]
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: path
          in: path
          required: true
          schema:
            type: string
        - name: uuid
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: integer
          in: header
          required: false
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParameterModel'
...
components:
  schemas:
...
    ParameterModel:
      type: object
      properties:
        bool:
          type: string
        string:
          type: string
        number:
          type: number
        integer:
          type: integer
          format: int32
        uuid:
          type: string
          format: uuid
        date:
          type: string
          format: date
        dateTime:
          type: string
          format: date-time
        a:
          type: array
          items:
            type: string
        b:
          type: array
          items:
            type: string
            format: uuid
        c:
          type: array
          items:
            type: integer
            format: int32